feat: 챌린지 진행 중 실시간 운동 데이터 수집, 랭킹 반영, 서버 간 동기화, 카프카 이벤트

+++ backend/runnity-websocket/src/main/java/com/runnity/websocket/dto/kafka/KafkaStreamEvent.java
@@ -33,7 +33,7 @@ public record KafkaStreamEvent(
     String nickname,
     String profileImage,
     @PositiveOrZero Double distance,
-    @PositiveOrZero Double pace,
+    @PositiveOrZero Integer pace,
     Integer ranking,
     Boolean isBroadcast,
     String reason,
